+++ experimental/cpp-serializer/include/types.h
@@ -3,6 +3,8 @@
 /******************************************************************************/
 
 #include <map>
+#include <list>
+#include <vector>
 #include <memory>
 
 /******************************************************************************/
@@ -13,6 +15,11 @@ using uPtr = std::unique_ptr<T>;
 template<typename T>
 using sPtr = std::shared_ptr<T>;
 
+template<typename T>
+using sVec = std::vector<T>;
+
+template<typename T>
+using sList = std::list<T>;
 
 template<typename T>
 using upStrMap_t = std::map<std::string, uPtr<T>>;
